return of the manufactory
After a long struggle against the crashes, the devestated developers called in the mighty hero of "manufactory", who quickly and unmercyfully purged the crashes. only a few pockets of crash resistance hide in the air nomad decisions now, but they will soon be overwhelmed by a supperior code bombardment. this final victory will long be remembered